+++ src/components/App.jsx
@@ -2,7 +2,7 @@ import { Route, Routes, BrowserRouter } from 'react-router-dom';
 import { InitialElement } from './InitialElement';
 import MainLayout from './MainLayout';
 import AccountPage from './../pages/AccountPage';
-import RegisterPage from 'pages/RegisterPage';
+import RegisterPage from 'pages/RegisterPage/RegisterPage';
 
 export const App = () => {
   return (
