+++ src/pages/RegisterPage/RegisterPage.jsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from '../../components/RegisterForm';
+import { RegisterForm } from '../../components/RegisterForm/RegisterForm';
 
 const RegisterPage = () => {
   return <RegisterForm />;
